+++ neurox/device/__init__.py
@@ -1,7 +1,7 @@
 """NeuroX device models."""
 
-from .nmos import NMOS, NMOSDCOP, NMOSConfig, NMOSPolicy, NMOSSnapshot
-from .rram import RRAM, RRAMDCOP, RRAMConfig, RRAMPolicy, RRAMSnapshot
+from .nmos import NMOS, NMOSDCOP, NMOSConfig, NMOSPolicy, NMOSSnap
+from .rram import RRAM, RRAMDCOP, RRAMConfig, RRAMPolicy, RRAMSnap
 from .selector import Selector, SelectorConfig, SelectorPolicy
 
 __all__ = [
@@ -10,13 +10,13 @@ __all__ = [
     "NMOSConfig",
     "NMOSDCOP",
     "NMOSPolicy",
-    "NMOSSnapshot",
+    "NMOSSnap",
     # --- RRAM ---
     "RRAM",
     "RRAMConfig",
     "RRAMDCOP",
     "RRAMPolicy",
-    "RRAMSnapshot",
+    "RRAMSnap",
     # --- Selector ---
     "Selector",
     "SelectorConfig",